move folders so that they align with expectations
