+++ array/CMakeLists.txt
@@ -2,4 +2,7 @@
 # Library
 ################################################################################
 add_library(consteig_array INTERFACE)
-target_include_directories(consteig_array INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
+target_include_directories(consteig_array INTERFACE
+    ${CMAKE_SOURCE_DIR}/include/consteig/array
+    ${CMAKE_SOURCE_DIR}/include
+)
