+++ eigen/CMakeLists.txt
@@ -2,4 +2,7 @@
 # Library
 ################################################################################
 add_library(consteig_eigen INTERFACE)
-target_include_directories(consteig_eigen INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
+target_include_directories(consteig_eigen INTERFACE
+    ${CMAKE_SOURCE_DIR}/include/consteig/eigen
+    ${CMAKE_SOURCE_DIR}/include
+)
