+++ include/consteig/consteig_options.hpp
