+++ include/consteig/math/constmath.hpp
