+++ include/consteig/math/functions/trig.hpp
