+++ include/consteig/math/functions/utilities.hpp
