+++ include/consteig/matrix/operations.hpp
