+++ math/CMakeLists.txt
@@ -2,4 +2,7 @@
 # Library
 ################################################################################
 add_library(consteig_math INTERFACE)
-target_include_directories(consteig_math INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
+target_include_directories(consteig_math INTERFACE
+    ${CMAKE_SOURCE_DIR}/include/consteig/math
+    ${CMAKE_SOURCE_DIR}/include
+)
